Move reduce impls into kernels::reduce
Technically not all of these are reduce ops.
That will have to wait for another day.
